refactor(dashboard-core): split selector ownership
