+++ packages/dashboard-core/test/unit/selectors/dashboardSessionRows.test.ts
@@ -4,6 +4,7 @@ import {
   sessionForWorktreeRow,
   sessionRowDisplayTitle,
 } from "../../../src/selectors/dashboardSessionRows.js";
+import { createInitialTuiState } from "../../../src/state/screen.js";
 import { createDashboardSnapshot, createExternalAgentSnapshot } from "../../fixtures/snapshots.js";
 
 const STATUS_DISPLAYS = {
@@ -85,6 +86,52 @@ describe("dashboard session rows", () => {
         },
       }),
     ).toBe("Optimistic readable title");
+  });
+
+  it("resolves canonical row titles", () => {
+    const snapshot = createDashboardSnapshot();
+    const session = snapshot.sessions.find((candidate) => candidate.id === "ses_wt_web_idle");
+    const worktree = snapshot.rows.find((candidate) => candidate.id === "wt_web_idle");
+    if (session === undefined || worktree === undefined) throw new Error("missing fixture session");
+    const titledWorktree = { ...worktree, title: "Readable feature task" };
+
+    expect(
+      sessionRowDisplayTitle(
+        { session, worktree: titledWorktree },
+        createInitialTuiState().localRows,
+      ),
+    ).toBe("Readable feature task");
+  });
+
+  it("retains a canonical no-agent title through pending launch state", () => {
+    const snapshot = createDashboardSnapshot();
+    const worktree = snapshot.rows.find((candidate) => candidate.id === "wt_web_no_agent");
+    const sourceSession = snapshot.sessions[0];
+    if (worktree === undefined || sourceSession === undefined) {
+      throw new Error("missing retained no-agent fixture inputs");
+    }
+    const session = { ...sourceSession, id: "ses_retained_no_agent", worktreeId: worktree.id };
+
+    expect(
+      sessionRowDisplayTitle(
+        { session, worktree: { ...worktree, title: "Durable no-agent workspace" } },
+        {
+          pendingCreate: [],
+          failedCreate: [],
+          pendingRemove: [],
+          pendingStart: [
+            {
+              localId: "start-retained",
+              operation: "resumeAgent",
+              projectId: worktree.projectId,
+              worktreeId: worktree.id,
+              branch: worktree.branch,
+              createdAt: "2026-05-31T12:00:00.000Z",
+            },
+          ],
+        },
+      ),
+    ).toBe("Durable no-agent workspace");
   });
 
   it("prefers external run identity over retained worktree membership", () => {
